Fix aspect ratio, add resizing and fullscreen toggle

+++ AllianceEngine/Scripts/Components/Camera.cs
@@ -13,7 +13,7 @@ namespace AllianceEngine
 
         public Matrix4x4 View => Matrix4x4.CreateLookAt(parent.Transform.Position, parent.Transform.Position
                                                                                    + parent.Transform.Forward, parent.Transform.Up);
-        public Matrix4x4 Projection => Matrix4x4.CreatePerspectiveFieldOfView(MathHelper.DegreesToRadians * CameraZoom, Program.Width / Program.Height, NearPlaneDistance, FarPlaneDistance);
+        public Matrix4x4 Projection => Matrix4x4.CreatePerspectiveFieldOfView(MathHelper.DegreesToRadians * CameraZoom, Program.Width / (float) Program.Height, NearPlaneDistance, FarPlaneDistance);
 
 
         public static void SwitchMainCamera(Camera camera)
